Add MSS54HP Validator
This builds on the handmade0octopus/Gauge.S-sd-updater repo and provides
a way to validate communication with the MSS54HP. This is set up to be
used as a PlatformIO project and flashed onto Gauge.S by following the
instructions in the README.

Once flashed, this will print out the TPS value via the serial console.
Actuating the throttle pedal should make the values change.